+++ etsin_finder/authorization.py
@@ -17,7 +17,7 @@ from etsin_finder.cr_service import (
     is_catalog_record_owner
 )
 
-from etsin_finder.finder import app
+from etsin_finder.app import app
 from etsin_finder import rems_service
 from etsin_finder.utils import tz_now_is_later_than_timestamp_str, remove_keys_recursively, leave_keys_in_dict
 from etsin_finder.constants import ACCESS_TYPES, DATA_CATALOG_IDENTIFIERS
