+++ etsin_finder/cr_service.py
@@ -9,7 +9,7 @@
 
 import requests
 
-from etsin_finder.finder import app
+from etsin_finder.app import app
 from etsin_finder.app_config import get_metax_api_config
 from etsin_finder.utils import json_or_empty, FlaskService, format_url
 from etsin_finder.constants import ACCESS_TYPES
