+++ etsin_finder/download_service.py
@@ -12,7 +12,7 @@ import requests
 from urllib import parse
 
 from etsin_finder.app_config import get_download_api_config
-from etsin_finder.finder import app
+from etsin_finder.app import app
 from etsin_finder.utils import FlaskService, format_url
 
 log = app.logger
