+++ etsin_finder/finder.py
@@ -5,61 +5,10 @@
 # :author: CSC - IT Center for Science Ltd., Espoo Finland <servicedesk@csc.fi>
 # :license: MIT
 
-"""Main app initialization file"""
+"""Main finder initialization file"""
 
-import os
-import logging
-import logging.config
-
-from flask import Flask
-from flask_mail import Mail
 from flask_restful import Api
-from flask.logging import default_handler
-
-from etsin_finder.app_config import get_app_config
-from etsin_finder.cache import CatalogRecordCache, RemsCache
-from etsin_finder.utils import executing_travis, get_log_config
-from etsin_finder.converters import IdentifierConverter
-
-
-def create_app():
-    """Create flask app
-
-    Returns:
-        objetc: The flask.Flask app instance object.
-
-    """
-    is_testing = bool(os.environ.get('TESTING', False))
-    app = Flask(__name__, template_folder="./frontend/build")
-    app.config.update(get_app_config(is_testing))
-    if not app.testing and not executing_travis():
-        _setup_app_logging(app)
-    if not executing_travis():
-        app.config.update({'SAML_PATH': '/home/etsin-user'})
-    app.mail = Mail(app)
-    app.cr_cache = CatalogRecordCache(app)
-    app.rems_cache = RemsCache(app)
-    app.url_map.converters['id'] = IdentifierConverter
-
-    return app
-
-
-def _setup_app_logging(app):
-    """Setup app logging
-
-    Load the logging configurations for the flask app.
-
-    Args:
-        app (object): flask.Flask object instance.
-
-    """
-    log_file_path = app.config.get('APP_LOG_PATH', None)
-    log_lvl = app.config.get('APP_LOG_LEVEL', 'INFO')
-    config = get_log_config(log_file_path, log_lvl)
-    if config:
-        logging.config.dictConfig(config)
-    else:
-        app.logger.error('Logging not correctly set up due to missing app log path configuration')
+from etsin_finder.app import app
 
 
 def add_restful_resources(app):
@@ -151,8 +100,8 @@ def add_restful_resources(app):
     api.add_resource(REMSApplyForPermission, '/api/rems/<id:cr_id>')
 
 
-app = create_app()
 add_restful_resources(app)
+app = app
 import etsin_finder.views
 
 if __name__ == "__main__":
