+++ etsin_finder/qvain_light_resources_v2.py
@@ -12,7 +12,7 @@ from flask import request
 from flask_restful import reqparse, Resource
 
 from etsin_finder import authentication
-from etsin_finder.finder import app
+from etsin_finder.app import app
 from etsin_finder.utils import (
     slice_array_on_limit,
     datetime_to_header
