+++ etsin_finder/qvain_light_rpc_v2.py
@@ -16,7 +16,7 @@ from etsin_finder.qvain_light_service_v2 import (
     merge_draft,
     create_draft
 )
-from etsin_finder.finder import app
+from etsin_finder.app import app
 from etsin_finder.log_utils import log_request
 from etsin_finder.qvain_light_utils_v2 import check_dataset_creator
