+++ etsin_finder/qvain_light_utils.py
@@ -9,13 +9,18 @@ from etsin_finder.constants import DATA_CATALOG_IDENTIFIERS, ACCESS_TYPES
 from etsin_finder.cr_service import (
     get_catalog_record,
 )
-from etsin_finder.finder import app
+from etsin_finder.app import app
 from etsin_finder.authentication import (
-    get_user_ida_groups, get_user_csc_name, get_user_email, get_user_firstname, get_user_lastname,
+    get_user_ida_groups,
+    get_user_csc_name,
+    get_user_email,
+    get_user_firstname,
+    get_user_lastname,
     is_authenticated
 )
 
 log = app.logger
+
 
 def clean_empty_keyvalues_from_dict(d):
     """Cleans all key value pairs from the object that have empty values, like [], {} and ''.
