+++ etsin_finder/request_utils.py
@@ -9,10 +9,11 @@
 
 import requests
 
-from etsin_finder.finder import app
+from etsin_finder.app import app
 from etsin_finder.utils import json_or_text
 
 log = app.logger
+
 
 def make_request(request_func, *args, **kwargs):
     """
